fix: resolve remaining build errors across 5 files
err.rs: add missing description()/code() arms for AssetDecimalsMismatch, DuplicateMarketId, CumulativeExtensionCapHit, IllegalMarketStateTransition\noracles.rs: add type annotation for twap() result, add history_size/max_deviation_z_multiple to GlobalOracleValidationConfig default\nmarkets.rs: fix u32/u64 mismatch in total_extension_days.saturating_add\nupgrade_manager.rs: fix chain.get() expecting u32 (cast u64→u32)\nlib.rs: add history_size/max_deviation_z_multiple to config initializers

+++ contracts/predictify-hybrid/src/markets.rs
@@ -1467,7 +1467,7 @@ impl MarketStateManager {
         }
 
         // Track cumulative extension
-        market.total_extension_days = market.total_extension_days.saturating_add(extension_hours);
+        market.total_extension_days = market.total_extension_days.saturating_add(extension_hours as u32);
 
         Ok(())
     }
