+++ contracts/predictify-hybrid/src/upgrade_manager.rs
@@ -650,7 +650,7 @@ impl UpgradeManager {
         let zero_hash = BytesN::from_array(env, &[0u8; 32]);
 
         for i in 0..verify_count {
-            let record = chain.get(i).ok_or(Error::InvalidInput)?;
+            let record = chain.get(i as u32).ok_or(Error::InvalidInput)?;
 
             if i == 0 {
                 // First record: previous_wasm_hash must be zero for genesis
@@ -659,7 +659,7 @@ impl UpgradeManager {
                 }
             } else {
                 // Subsequent records: previous_wasm_hash must match previous record's new_wasm_hash
-                let prev_record = chain.get(i - 1).ok_or(Error::InvalidInput)?;
+                let prev_record = chain.get((i - 1) as u32).ok_or(Error::InvalidInput)?;
                 if record.previous_wasm_hash != prev_record.new_wasm_hash {
                     return Ok(false);
                 }
